Create basic Hello World page
My initial plan was to use [this tutorial](https://www.freecodecamp.org/news/how-i-set-up-a-real-world-project-with-go-and-vue/)
to make a basic Go/Vue site and then adapt that to the actual task. As
I've never used either technology before, I thought this would be a good
way to learn the stack and get familiar.

However, for some reason I cannot get the ``vue create`` command to work
at all. It appears to run successfully, and even prints the next
commands to run the server, but no files are created by it.

Rather than spend any more time on this, I'm going to approach this as
if I was just a backend developer and focus on making a Go project which
could plug into a Vue frontend. Hopefully I can come back to Vue later.