Refactor message to new struct
This was more just an experiment to learn how structs, interfaces and
constructors work in Go.

My initial thought behind this was that actual application probably
shouldn't all live in one file. In OOP it would be standard to have one
file per class. However, it seems that Go discourages this idea, and
instead has a "one concept per file" approach instead. They also seem to
discourage MVC, which was my first thought when I read the project.

I'm trying to adapt Go's best practices as I go, so I've put the interface
and implementation in one file. I think this makes sense as a "concept".
The server file is handling requests, which seems different enough to
merit its own file.