Create struct to hold assets
Also include a stub method to return the particular assets for the test.
In reality I would expect these to come from a database or another API.

I'm assuming from the example instructions that the assets have values
and quantities. I've made up some values for these which match the %s
given.

I'm also assuming that the %s given are not necessarily stored
themselves. Partly this is because I thought it would be more common for
a customer to buy or sell particular assets, than aim for a certain %.
Mainly though, I just don't like storing values that can otherwise be
derivied (unless efficiency is a major concern, which I don't think it
is here). This helps avoid contradictions with mismatched values.

Finally, I am assuming that assets can have fractional (floats)
quantities. This is based on a very brief bit of research. This also
helps a lot of the maths.