Create interface to manage a portfolio
This has a single function, to produce a message to withdraw funds. I
am assuming this class should only produce the messages to manage the
funds and not actually change the asset amounts. Presumably this is
the job of the message receiver.

The WriteInstruction function can easily be extended to the other
instruction types. While it would be trivial to add these other
functions, I did not as this was not part of the spec.

I considered creating another function to write each line of the
instruction as there's some repetition. However, minor differences (e.g.
quotes) made me think it wasn't worth it.

Usually I avoid having stings that aren't const-style variables, but
I've found they hinder readability when it comes to single characters
like quotes or brackets.

I have assumed that when withdrawing the relative % of stocks needs to
remain the same.