Refactor server and main to serve portfolio api
At this point I realised I forgot to commit main.go before. It already
existed to pass the HomePage struct to the server using the same
pattern. This is to decouple the server from the interface. homePage.go
can now be removed as it's no longer needed. Instead, have main.go get
the assets from the stub method and pass those to the portfolio, and the
portfolio to server.

To begin, simply have the server send a Withdraw request of £10. I tried
a few different amounts before commiting to test this.

If I had more time, the next step would be to change this to a post
request with a withdrawl amount in the parameters. After that, I would
probably need to change the output to something Vue can read easier.